Implement colored texturing

+++ raycaster_data.h
@@ -2,4 +2,4 @@
 
 extern const uint8_t LOOKUP_TBL g_map[128];
 
-extern const uint8_t LOOKUP_TBL g_texture8[4096];
+extern const uint32_t LOOKUP_TBL g_texture8[4096];
